fix: decrypted filenames after each refresh, moved filename decryption function to files store. Added 'await' for async functions to prevent conflicts.
